refactor(opcache): move SharedMemoryException and add SharedMemory facade
SharedMemoryException moves from ZEngine\Memory into ZEngine\OpCache
alongside the rest of the opcache surface (it still extends
ReflectionException, so existing catch blocks keep working), and a new
ZEngine\OpCache\SharedMemory facade centralizes immutability detection by
delegating to the reflection wrappers that own the engine flags. Fixes the
docs that still referenced the deleted ZEngine\Memory\SharedMemory class.
