+++ docs/hot-swap.md
@@ -111,8 +111,10 @@ Opcache publishes cached scripts from shared memory and marks their class
 entries and functions `ZEND_ACC_IMMUTABLE`. SHM is visible to every worker
 process, so z-engine never writes it and never frees it. The wrappers expose
 the detection as `ReflectionClass::isImmutable()` and
-`ReflectionFunction`/`ReflectionMethod::isImmutable()`; the shared detection
-and copy-out mechanics both delegate to live in `ZEngine\Memory\SharedMemory`.
+`ReflectionFunction`/`ReflectionMethod::isImmutable()`, surfaced together on
+the `ZEngine\OpCache\SharedMemory` facade; the copy-out of an immutable global
+function happens inside `ReflectionFunction::redefine()`
+(`FunctionLikeTrait::copyOutOfSharedMemory()`).
 
 | Target | Behaviour |
 |--------|-----------|
