+++ docs/long-running.md
@@ -129,7 +129,7 @@ The debug-build leak gate treats the following as expected, by design:
 | One live libffi trampoline per installed hook | owned by ext/ffi, freed by its RSHUTDOWN |
 | One `zend_object_iterator_funcs` vtable for the get-iterator bridge (`IteratorBridge`) | live engine iterators dereference `->funcs` for their whole lifetime; the block is a malloc-backed process-wide singleton filled with libffi trampolines (trampolines themselves owned by ext/ffi). The bridge registers in the Core hook registry, so `Core::shutdown()` neutralizes surviving iterators (drops their cached current-value reference, swaps their handlers to `std_object_handlers`) while trampolines are still alive, and `Core::reinstallHooks()` re-mints the vtable for cycling SAPIs |
 | Closures immortalized by `ReflectionClass::addMethod()` | the method table references the closure body for the rest of the request |
-| Copied-out shared-memory function containers (`SharedMemory::copyOutFunctionEntry()`) and the SHM originals they replace | opcache SHM is never written or freed; the writable `zend_function` container is malloc-backed (the engine's table destructor releases the body but never frees user function containers); bounded to one container per copied-out function |
+| Copied-out shared-memory function containers (`FunctionLikeTrait::copyOutOfSharedMemory()`) and the SHM originals they replace | opcache SHM is never written or freed; the writable `zend_function` container is malloc-backed (the engine's table destructor releases the body but never frees user function containers); bounded to one container per copied-out function |
 | Added-method `zend_function` and added-constant `zend_class_constant` containers (`ClassDelta::apply()`) | published table buckets reference them until the class dies; the engine releases the bodies/values they carry but never frees the containers themselves (it assumes arena storage); malloc-backed, bounded to one per added member |
 | Removed-method structures and their bodies (`ClassDelta::apply()`) | warmed-up inline caches and subclass method buckets may still point at the unpublished entry, so neither the structure nor its opcodes can be freed; bounded to one body per removed method |
 | Previous function body of an in-flight `redefine()` (the entry was on the VM call stack during the swap) | the running frame still executes the old opcodes; the swap keeps the body allocated instead of freeing memory under the VM; bounded to one body per such self-redefinition |
