+++ tests/HotSwap/scripts/opcache-matrix.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
 use ReflectionMethod as NativeReflectionMethod;
 use ZEngine\Core;
 use ZEngine\HotSwap\HotSwap;
-use ZEngine\Memory\SharedMemoryException;
+use ZEngine\OpCache\SharedMemoryException;
 use ZEngine\Reflection\ReflectionClass;
 use ZEngine\Reflection\ReflectionFunction;
 use ZEngine\Reflection\ReflectionMethod;
